refactor: move wire to di package and provide oAuthConfig
