+++ internal/di/wire.go
@@ -1,13 +1,14 @@
 //go:build wireinject
 // +build wireinject
 
-package framework
+package di
 
 import (
 	"github.com/d0lim/turnstile/internal/adapters/in/api"
 	"github.com/d0lim/turnstile/internal/adapters/out/db"
 	"github.com/d0lim/turnstile/internal/adapters/out/db/ent"
 	"github.com/d0lim/turnstile/internal/core/ports/in/usecase"
+	"github.com/d0lim/turnstile/internal/framework"
 	"github.com/d0lim/turnstile/internal/framework/config"
 	"github.com/gofiber/fiber/v2"
 	"github.com/google/wire"
@@ -16,11 +17,12 @@ import (
 func InitializeApp() (*fiber.App, error) {
 	wire.Build(
 		ent.NewClient,
-		config.NewSessionStore,
+		config.NewSessionConfig,
+		config.NewOAuthConfig,
 		db.NewUserRepository,
 		usecase.NewUserUsecase,
 		api.NewUserHandler,
-		NewApp,
+		framework.NewApp,
 	)
 
 	return &fiber.App{}, nil
